Update AvailabilityManager so adding an observer will always lead to an immediate callback (important when there are no other nodes or all availability info is otherwise known).
Add UI elements for download UI to display availability.

+++ lib-database-entities/src/commonMain/kotlin/com/ustadmobile/lib/db/entities/AvailabilityObserverItem.kt
@@ -18,7 +18,7 @@ import kotlinx.serialization.Serializable
 open class AvailabilityObserverItem {
 
     @PrimaryKey(autoGenerate = true)
-    var aoiId: Long = 0
+    var aoiId: Int = 0
 
     var aoiOriginalUrl: String? = null
